Streak
The streak was added, to avoid errors with the previous users, which didn't had such parameter a updateStreak and a calculateStreak methods were created.

Otherwise, it is a simple integer (set as so to avoid database problems - instead of a primitive int) that goes to zero if there wasn't any activities yesterday, and none today.